fix: remove method on new config class
